software/pyproject.toml
[project]
name = "current-sensor-reader"
version = "0.1.0"
description = "Software to reading serial data from custom current sensor"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "pyserial>=3.5",
]
